refactor: CommandMapper에서 관리하는 명령어를 다른 클래스들이 사용하도록 변경
