+++ src/main/java/command/End.java
@@ -1,10 +1,9 @@
 package command;
 
 import domain.game.ChessGame;
+import view.CommandMapper;
 
 public class End implements Command {
-
-    private static final String END_COMMAND = "end";
 
     private End() {
     }
@@ -15,7 +14,7 @@ public class End implements Command {
     }
 
     private static void validate(String command) {
-        if (!END_COMMAND.equals(command)) {
+        if (!CommandMapper.END.commandText().equals(command)) {
             throw new IllegalArgumentException("올바른 명령어를 입력해 주세요.");
         }
     }
