+++ src/main/java/command/Save.java
@@ -3,12 +3,11 @@ package command;
 import dao.BoardDao;
 import dao.TurnDao;
 import domain.game.ChessGame;
+import view.CommandMapper;
 
 import static view.OutputView.printSaveMessage;
 
 public class Save implements Command {
-
-    private static final String SAVE_COMMAND = "save";
 
     private Save() {
     }
@@ -19,7 +18,7 @@ public class Save implements Command {
     }
 
     private static void validate(String command) {
-        if (!SAVE_COMMAND.equals(command)) {
+        if (!CommandMapper.SAVE.commandText().equals(command)) {
             throw new IllegalArgumentException("올바른 명령어를 입력해 주세요.");
         }
     }
