+++ src/main/java/command/Start.java
@@ -1,10 +1,9 @@
 package command;
 
 import domain.game.ChessGame;
+import view.CommandMapper;
 
 public class Start implements Command {
-
-    private static final String START_COMMAND = "start";
 
     private Start() {
     }
@@ -15,7 +14,7 @@ public class Start implements Command {
     }
 
     private static void validate(String command) {
-        if (!START_COMMAND.equals(command)) {
+        if (!CommandMapper.START.commandText().equals(command)) {
             throw new IllegalArgumentException("올바른 명령어를 입력해 주세요.");
         }
     }
